pipeline/src/requirements.txt
selenium
PyYAML
webdriver-manager
numpy
pymongo
nltk
spacy
bs4
tensorflow
transformers